requirements.txt
torch
torchvision
scikit-learn
matplotlib
pandas
pillow
tqdm
seaborn
streamlit
opencv-python
numpy
